cmd/python-worker/requirements.txt
boto3==1.34.57
requests==2.31.0
# tensorflow==2.15.0
tensorflow-cpu==2.15.0
numpy==1.26.4
